CBB: 	* Une unsigned data types for dimensions/sizes 	  That's really what the code expects, and everything is already 	  bounds-checked on the Lua side.

+++ blitbuffer.h
@@ -48,88 +48,88 @@ typedef struct ColorRGB32 {
 } ColorRGB32;
 
 typedef struct BlitBuffer {
-    int w;
-    int pixel_stride;   // nb of pixels from the start of a line to the start of next line
-    int h;
-    int stride;         // nb of bytes from the start of a line to the start of next line
+    unsigned int w;
+    unsigned int pixel_stride;   // nb of pixels from the start of a line to the start of next line
+    unsigned int h;
+    size_t stride;         // nb of bytes from the start of a line to the start of next line
     uint8_t *data;
     uint8_t config;
 } BlitBuffer;
 
 typedef struct BlitBuffer8 {
-    int w;
-    int pixel_stride;
-    int h;
-    int stride;
+    unsigned int w;
+    unsigned int pixel_stride;
+    unsigned int h;
+    size_t stride;
     Color8 *data;
     uint8_t config;
 } BlitBuffer8;
 
 typedef struct BlitBuffer8A {
-    int w;
-    int pixel_stride;
-    int h;
-    int stride;
+    unsigned int w;
+    unsigned int pixel_stride;
+    unsigned int h;
+    size_t stride;
     Color8A *data;
     uint8_t config;
 } BlitBuffer8A;
 
 typedef struct BlitBufferRGB16 {
-    int w;
-    int pixel_stride;
-    int h;
-    int stride;
+    unsigned int w;
+    unsigned int pixel_stride;
+    unsigned int h;
+    size_t stride;
     ColorRGB16 *data;
     uint8_t config;
 } BlitBufferRGB16;
 
 typedef struct BlitBufferRGB24 {
-    int w;
-    int pixel_stride;
-    int h;
-    int stride;
+    unsigned int w;
+    unsigned int pixel_stride;
+    unsigned int h;
+    size_t stride;
     ColorRGB24 *data;
     uint8_t config;
 } BlitBufferRGB24;
 
 typedef struct BlitBufferRGB32 {
-    int w;
-    int pixel_stride;
-    int h;
-    int stride;
+    unsigned int w;
+    unsigned int pixel_stride;
+    unsigned int h;
+    size_t stride;
     ColorRGB32 *data;
     uint8_t config;
 } BlitBufferRGB32;
 
-void BB_fill_rect(BlitBuffer *bb, int x, int y, int w, int h, uint8_t v);
-void BB_blend_rect(BlitBuffer *bb, int x, int y, int w, int h, Color8A *color);
-void BB_invert_rect(BlitBuffer *bb, int x, int y, int w, int h);
+void BB_fill_rect(BlitBuffer *bb, unsigned int x, unsigned int y, unsigned int w, unsigned int h, uint8_t v);
+void BB_blend_rect(BlitBuffer *bb, unsigned int x, unsigned int y, unsigned int w, unsigned int h, Color8A *color);
+void BB_invert_rect(BlitBuffer *bb, unsigned int x, unsigned int y, unsigned int w, unsigned int h);
 void BB_blit_to_BB8(BlitBuffer *src, BlitBuffer *dst,
-                    int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
+                    unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
 void BB_dither_blit_to_BB8(BlitBuffer *src, BlitBuffer *dst,
-                    int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
+                    unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
 void BB_blit_to_BB8A(BlitBuffer *src, BlitBuffer *dst,
-                     int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
+                     unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
 void BB_blit_to_BB16(BlitBuffer *src, BlitBuffer *dst,
-                     int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
+                     unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
 void BB_blit_to_BB24(BlitBuffer *src, BlitBuffer *dst,
-                     int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
+                     unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
 void BB_blit_to_BB32(BlitBuffer *src, BlitBuffer *dst,
-                     int dest_x, int dest_y, int offs_x, int offs_y, int w, int h);
-void BB_blit_to(BlitBuffer *source, BlitBuffer *dest, int dest_x, int dest_y,
-                int offs_x, int offs_y, int w, int h);
-void BB_dither_blit_to(BlitBuffer *source, BlitBuffer *dest, int dest_x, int dest_y,
-                int offs_x, int offs_y, int w, int h);
-void BB_add_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                      int offs_x, int offs_y, int w, int h, uint8_t alpha);
-void BB_alpha_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                        int offs_x, int offs_y, int w, int h);
-void BB_pmulalpha_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                        int offs_x, int offs_y, int w, int h);
-void BB_dither_pmulalpha_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                        int offs_x, int offs_y, int w, int h);
-void BB_invert_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                         int offs_x, int offs_y, int w, int h);
-void BB_color_blit_from(BlitBuffer *dest, BlitBuffer *source, int dest_x, int dest_y,
-                        int offs_x, int offs_y, int w, int h, Color8A *color);
+                     unsigned int dest_x, unsigned int dest_y, unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_blit_to(BlitBuffer *source, BlitBuffer *dest, unsigned int dest_x, unsigned int dest_y,
+                unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_dither_blit_to(BlitBuffer *source, BlitBuffer *dest, unsigned int dest_x, unsigned int dest_y,
+                unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_add_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                      unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h, uint8_t alpha);
+void BB_alpha_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                        unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_pmulalpha_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                        unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_dither_pmulalpha_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                        unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_invert_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                         unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h);
+void BB_color_blit_from(BlitBuffer *dest, BlitBuffer *source, unsigned int dest_x, unsigned int dest_y,
+                        unsigned int offs_x, unsigned int offs_y, unsigned int w, unsigned int h, Color8A *color);
 #endif
